move example data to separate dir
