+++ plugins/cost-insights/src/example/alerts/index.ts
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-export { AlertAcceptForm } from './AlertAcceptForm';
-export { AlertSnoozeForm } from './AlertSnoozeForm';
-export { AlertDismissForm } from './AlertDismissForm';
+export { KubernetesMigrationAlert } from './KubernetesMigrationAlert';
+export type { KubernetesMigrationApi } from './KubernetesMigrationAlert';
+export type { KubernetesMigrationData } from './KubernetesMigrationAlert';
