+++ plugins/cost-insights/src/example/components/index.ts
@@ -14,6 +14,4 @@
  * limitations under the License.
  */
 
-export { AlertAcceptForm } from './AlertAcceptForm';
-export { AlertSnoozeForm } from './AlertSnoozeForm';
-export { AlertDismissForm } from './AlertDismissForm';
+export * from './KubernetesMigrationAlertCard';
