+++ plugins/cost-insights/src/example/index.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
-export { AlertAcceptForm } from './AlertAcceptForm';
-export { AlertSnoozeForm } from './AlertSnoozeForm';
-export { AlertDismissForm } from './AlertDismissForm';
+export * from './alerts';
+export * from './components';
+export * from './forms';
+export * from './client';
