+++ plugins/cost-insights/src/index.ts
@@ -21,9 +21,9 @@ export {
   CostInsightsProjectGrowthInstructionsPage,
   CostInsightsLabelDataflowInstructionsPage,
 } from './plugin';
-export * from './client';
-export * from './api';
 export { ProjectGrowthAlert, UnlabeledDataflowAlert } from './alerts';
+export { ExampleCostInsightsClient } from './example';
+export * from './api';
 export * from './components';
 export { useCurrency } from './hooks';
 export * from './types';
